Fix #2 and add stdin
* Fixes https://github.com/fernandrone/linelint/issues/2. The issue is
that the "autofix" command line flag was overriding the Config object.
So even if the config had autofix: true, later in the main function the
flag value would override this configuration. It didn't help that the
default bool value for the flag was 'false'.

* Refactor the main code at linelint.go, allowing for input from STDIN.
This took some inspiration from autopep8, in which to read from STDIN
requires the "-" positional argument (any other positional argument
is interpreted as a list of files/dirs to lint, and an "empty" list is
interpreted as "lint the current directory tree"). This also makes it
much easier to test the linting from the main entrypoint.

* Add a couple tests as Example in linelint_test.go, using the new
"read from stdin" functionality.